==> FIFA_WC_Analysis/1.get_groupstage_tables.ipynb <==
import pandas as pd
from string import ascii_uppercase as alphabet
import pickle

#extracting all tables in website
all_tables = pd.read_html('https://web.archive.org/web/20221115040351/https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')


dict_table = {}

for letter, i in zip(alphabet,range(12,68,7)):
    df=all_tables[i]
    df.rename(columns={df.columns[1]:'Team'}, inplace=True)
    df.pop('Qualification')
    dict_table[f'Group {letter}'] = df

# df = all_tables[12]
# df = all_tables[19]
# df = all_tables[61]

dict_table.keys()

dict_table['Group A']

# Pickle library
the pickle library in Python is used for serializing and deserializing Python objects. It allows you to store complex Python objects, such as lists, dictionaries, and custom classes, in a binary format that can be saved to disk, transmitted over a network, or stored in memory. In this example, the pickle.dump method is used to save the data object to a file, and the pickle.load method is used to load the object from the file. The wb and rb modes are used to indicate that the file should be opened in binary mode for writing and reading, respectively.

# Upload (..verify if uploaded correctly)
with open('dict_table', 'wb') as output:
    pickle.dump(dict_table, output)
