==> FIFA_WC_Analysis/1a.get_results_&_fixtures.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import requests



# data to scraper -> 1930-2018
years = [1930, 1934, 1938, 1950, 1954, 1958, 1962, 1966, 1970, 1974,
         1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014,
         2018]

def get_matches(year):
    web = f'https://en.wikipedia.org/wiki/{year}_FIFA_World_Cup'
    response = requests.get(web)
    content = response.text
    soup = BeautifulSoup(content, 'lxml')
    matches = soup.find_all('div', class_='footballbox')

    home = []
    score = []
    away = []

    for match in matches:
        home.append(match.find('th', class_='fhome').get_text())
        score.append(match.find('th', class_='fscore').get_text())
        away.append(match.find('th', class_='faway').get_text())

    dict_football = {'home': home, 'score': score, 'away': away}
    df_football = pd.DataFrame(dict_football)
    df_football['year'] = year
    return df_football

#fixtures
web22 = f'https://web.archive.org/web/20221115040351/https://en.wikipedia.org/wiki/2022_FIFA_World_Cup'
response22 = requests.get(web22)
content22 = response22.text
soup22 = BeautifulSoup(content22, 'lxml')
matches_2022 = soup22.find_all('div', class_='footballbox')

home_2022 = []
score_2022= []
away_2022= []

for match22 in matches_2022:
    home_2022.append(match22.find('th', class_='fhome').get_text())
    score_2022.append(match22.find('th', class_='fscore').get_text())
    away_2022.append(match22.find('th', class_='faway').get_text())

dict_football2022 = {'home': home_2022, 'score': score_2022, 'away': away_2022}
df_fixture = pd.DataFrame(dict_football2022)


df_fixture

# results: historical data
fifa = [get_matches(year) for year in years]
df_fifa = pd.concat(fifa, ignore_index=True)
df_fifa.to_csv("fifa_worldcup_historical_data.csv", index=False)

# fixture
df_fixture = get_matches(2022)
df_fixture.to_csv('fifa_worldcup_fixture.csv', index=False)
